This function maps query data against the AIT (Shiny) reference taxonomy using hierarchical mapping.

This function maps query data against the AIT (Shiny) reference taxonomy using hierarchical mapping.

get_mapmycells_results(
  query_data_output_path,
  extended_result_path,
  precomp_stats_output_path,
  query_markers_output_path,
  mapping_params_list
)

Arguments

query_data_output_path

Full file path to the query data, a logCPM normalized matrix to be annotated that's saved as AnnData H5AD file.

extended_result_path

Full file path and name where the original mapping results will be saved.

precomp_stats_output_path

Full file path to the precompute stats HDF5 file.

query_markers_output_path

Full file path to the query markers JSON file.

chunk_size

Number of rows each worker process should load at a time from the query dataset.

n_processors

Number of independent worker processes to spin up.

normalization

Normalization of the h5ad files; must be either 'raw' or 'log2CPM'.

Value

resuls